rbs.rake: Skip to add tasks if rbs_rails is not able to load
rbs_rails is usually used in development mode. So it may be not able to load in the production mode. At present, the generated rbs.rake loads `rbs_rails/rake_task`. But it might fail in many projects.

This skips to add tasks if rbs_rails is not able to load.