CI: Add GitHub Actions workflow for releasing gem
This allows releasing a new gem package by changing the version.rb.  It
helps to simplify and automate the release process.  It also helps not
to grant permissions on rubygems.org to the co-maintainers  (See
[Trusted publishing](https://guides.rubygems.org/trusted-publishing/)).

This workflow will be triggered when version.rb has been changed on the
main branch.  And then, GitHub Action will create a new git tag with
the version and release the HEAD of the main branch as a new gem.

Thanks to @sinsoku!