feat: Add command line interface for rbs_rails
https://github.com/pocke/rbs_rails/pull/296 added `rbs_rails:prepare`
task to install inspector modules for Active Record models.  It is
required to run before loading Rails application.  It is an incompatible
change and hard to understand for users.

To avoid this, this introduces a new command `rbs_rails` to extract
types from Rails app in the independent process with Rake.  It allows to
install inspectors before loading Rails application easily.

Note: Sadly, the new command does not support the `ignore_model_if`
option.