cli: Implement `rbs_rails server` command
To make RBS generation faster to extract signatures from the Ruby on
Rails application, this provides the rbs_rails LSP server.  It generates
signatures on demand if the IDE lets the rbs_rails LSP server know the
changes to Ruby files.

It preloads the Rails application on start and reloads it when the
source code is changed. It improves the performance of signature
generation rather than running the rbs_rails command