cli: Implement `rbs_rails server` command (prototype)
To make RBS generation faster to extract signatures from the Ruby on
Rails application, this provides the rbs_rails server.  It generates
signatures on demand if the IDE lets the rbs_rails server know the
changes to Ruby files.

The rbs_rails server preloads the Rails application on start and
reloads it when the source code is changed. It improves the performance
of signature generation rather than running the rbs_rails command

The server accepts a POST request with a JSON body containing the "path"
property, which is the path of the Ruby file from the Rails root.

```
curl -X POST http://localhost:8080 -d '{"path": "/app/models/user.rb"}'
```