Check database migrations before generating types for models
To ensure consistency between the database schema and the generated
types, this adds a new feature that rbs_rails stops to generate types
for models if the database is not migrated to the latest version.

This feature can be configured via `check_db_migrations` settings.

Thanks to @sinsoku

Example:

```
$ bundle exec rbs_rails models
Error:

Migrations are pending. To resolve this issue, run:

        bin/rails db:migrate

You have 1 pending migration:

db/migrate/20250812173612_add_email_to_users.rb
```