active_record: Support `#id_value` alias for `#id` column
Since Rails 7.1, ActiveRecord provides `#id_value` alias for the `#id`
column.  It will be defined if the model has `id` column.

Note: The alias definition is not stored in `attribute_aliases`.  It is
generated on the instantiation of the model automatically.  Therefore,
we need to give special treatment to generate `#id_value` alias like
ActiveRecord does.

refs: https://github.com/rails/rails/pull/48930